product-service/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.0
dnspython==2.7.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pika==1.3.2
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
uritemplate==4.2.0
psycopg2-binary
drf-yasg2
django-prometheus
pillow
